requirements.txt
flask
python-dotenv
psycopg2-binary
flask_jwt_extended
Flask-SQLAlchemy
flask_marshmallow
marshmallow-sqlalchemy
flask_restful
werkzeug
sqlalchemy-json
gurobipy == 9.5.2
numpy